Fix warning about comparision of signed and unsigned ints
git-svn-id: https://crosswire.org/svn/sword/trunk@1792 bcd7d363-81e1-0310-97ec-a550e20fc99c

+++ include/swbuf.h
@@ -48,7 +48,7 @@ class SWDLLEXPORT SWBuf {
 	static char junkBuf[JUNKBUFSIZE];
 
 	inline void assureMore(size_t pastEnd) {
-		if (endAlloc-end < pastEnd) {
+		if (size_t(endAlloc-end) < pastEnd) {
 			assureSize(allocSize + pastEnd);
 		}
 	}
